requirements.txt
fastapi
uvicorn
fastapi-authlib
python-dotenv
sqlalchemy
python-decouple